python: restructure python bindings
Problem: The current structure of the Python bindings
could cause problems when running tests on python3.6.
/src/bindings/python/flux didn't appear to be a module
because the module loading  system doesn't see an __init__.py.
It could also end up breaking the import of other flux
modules from the system path after an __init__.py was
added.

Restructure the directory that holds the Python bindings
to be a namespace package called fluxacct. This namespace
will then hold a directory called accounting, which hosts
the current Python source code files.

Move the front-end user command into a separate directory
under src/, called cmd/.

Fixes #105
