+++ src/bindings/python/fluxacct/accounting/test/test_bank_subcommands.py
@@ -14,9 +14,9 @@ import os
 import sqlite3
 import pandas as pd
 
-from flux.accounting import accounting_cli_functions as aclif
-from flux.accounting import create_db as c
-from flux.accounting import print_hierarchy as p
+from fluxacct.accounting import accounting_cli_functions as aclif
+from fluxacct.accounting import create_db as c
+from fluxacct.accounting import print_hierarchy as p
 
 
 class TestAccountingCLI(unittest.TestCase):
