+++ src/bindings/python/fluxacct/accounting/test/test_job_archive_interface.py
@@ -18,9 +18,9 @@ import pandas as pd
 
 from unittest import mock
 
-from flux.accounting import job_archive_interface as jobs
-from flux.accounting import create_db as c
-from flux.accounting import accounting_cli_functions as aclif
+from fluxacct.accounting import job_archive_interface as jobs
+from fluxacct.accounting import create_db as c
+from fluxacct.accounting import accounting_cli_functions as aclif
 
 
 class TestAccountingCLI(unittest.TestCase):
